requirements.txt
matplotlib
numpy
cv2-tools